# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the VirtualYou Personal API server. You can find out more about this backend JSON server at [https://github.com/dlwhitehurst/peeps](https://github.com/dlwhitehurst/personal) ",
    "version": "0.0.0",
    "title": "Personal",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3006",
  "basePath": "personal/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/peeps/{peepId}": {
      "get": {
        "tags": [
          "peep resource"
        ],
        "summary": "Find Peep by ID",
        "description": "Returns a single Peep",
        "operationId": "getPeepById",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "peepId",
            "in": "path",
            "description": "ID of Peep to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Peep"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "peep resource"
        ],
        "summary": "Update an existing Peep",
        "description": "",
        "operationId": "updatePeepById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "peepId",
            "in": "path",
            "description": "ID of Peep to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Validation Exception"
          }
        }
      },
      "delete": {
        "tags": [
          "peep resource"
        ],
        "summary": "Deletes a Peep by ID",
        "description": "",
        "operationId": "deletePeep",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "peepId",
            "in": "path",
            "description": "ID of Peep to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid ID Supplied"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/owner/peeps/{peepId}": {
      "get": {
        "tags": [
          "peep resource"
        ],
        "summary": "Find Peep by ID",
        "description": "Returns a single Peep",
        "operationId": "getPeepById",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "peepId",
            "in": "path",
            "description": "ID of Peep to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Peep"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "peep resource"
        ],
        "summary": "Update an existing Peep",
        "description": "",
        "operationId": "updatePeepById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "peepId",
            "in": "path",
            "description": "ID of Peep to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Validation Exception"
          }
        }
      },
      "delete": {
        "tags": [
          "peep resource"
        ],
        "summary": "Deletes a Peep by ID",
        "description": "",
        "operationId": "deletePeep",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "peepId",
            "in": "path",
            "description": "ID of Peep to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid ID Supplied"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Peep": {
      "type": "object",
      "required": [
        "name",
        "code"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "North Carolina"
        },
        "code": {
          "type": "string",
          "example": "NC"
        }
      }
    }
  },
...
}